Updates to extract and process windfield

+++ main-wind.sh
@@ -7,9 +7,9 @@
 
 # Variable configurations
 gpt=~/tools/esa-snap10.0/bin/gpt
-dataset=cimat
+dataset=tarso
 base_path=~/data/cimat/dataset-$dataset
-source=products
+source=sentinel1
 tiff=tiff
 dest=wind
 temp=temp4
@@ -27,7 +27,9 @@ for fname in $(cat $base_path/${source}_list${num}.txt); do
   echo "Processing $name in bash script"
   echo "Extension $ext"
   # Transfer product from siimon5 GRD product directory
-  scp -P 2235 manuelsuarez@siimon5.cimat.mx:~/data/cimat/dataset-${dataset}/$source/$fname $base_path/$source/$fname
+  if [ ! -f $base_path/$source/$fname ]; then
+    scp -P 2235 manuelsuarez@siimon5.cimat.mx:~/data/cimat/dataset-${dataset}/$source/$fname $base_path/$source/$fname
+  fi
   # Unzip product and make temp directory
   if [ $ext == "zip" ]; then
     if [ ! -d $base_path/$source/$name.SAFE ]; then
@@ -60,36 +62,40 @@ for fname in $(cat $base_path/${source}_list${num}.txt); do
   while [ ! -f $base_path/$temp/$name/wind_05.dim ]; do
     # Sleep
     echo "$base_path/$temp/$name/wind_05 result has not been created, waiting 2m..."
-    sleep 2m
+    sleep 1m
   done
   echo "$base_path/$temp/$name/wind_05 result has been created, wait 5m and continue processing..."
   # We wait for if the GPT pipeline has not been finished
-  sleep 5m
+  #sleep 1m
   # Run windfield estimation (step 6)
   if ! test -f $base_path/$temp/$name/wind_06.dim; then
     $gpt scripts/wind_06.xml -SsourceProduct=$base_path/$temp/$name/wind_05.dim -t $base_path/$temp/$name/wind_06.dim
   fi
-  #
   # Run slurm postprocessing steps (step 7) script (on C0 cpu nodes)
   sbatch run-wind.slurm $base_path $source $dest $temp $fname 02
+  # Terrain correction
+  #if ! test -f $base_path/$temp/$name/wind_07.dim; then
+  #  $gpt scripts/wind_07.xml -SsourceProduct=$base_path/$temp/$name/wind_06.dim -t $base_path/$temp/$name/wind_07.dim
+  #fi
   # Wait until step 7 is finished
   while [ ! -f $base_path/$temp/$name/wind_07.dim ]; do
-    # Sleep
+  #  # Sleep
     echo "$base_path/$temp/$name/wind_07 result has not been created, waiting 2m..."
-    sleep 2m
+    sleep 1m
   done
   echo "$base_path/$temp/$name/wind_07 result has been created, wait 5m and continue processing..."
   # We wait for if the GPT pipeline has not been finished
-  sleep 5m
+  sleep 1m
   # There is no extract vector data (SHP) on GPT pipeline so we need to read the CSV directly on the BEAM
   # data and processing with Geopandas and Rasterio to interpolate to Wind Field Image
   #
   # Transfer original TIFF image to get dimensions for interpolation
-  scp -P 2235 manuelsuarez@siimon5.cimat.mx:~/data/cimat/dataset-${dataset}/$tiff/$name.tif $base_path/$temp/$name.tif
-  srun python read-wind-csv.py
+  #scp -P 2235 manuelsuarez@siimon5.cimat.mx:~/data/cimat/dataset-${dataset}/$tiff/$name.tif $base_path/$temp/$name.tif
+  srun --partition=C0 --job-name=ExtractWindField --time=0 --mem=0 python extract-windfield.py --fname=$name
   # Move result to siimon5 (we are implementing an active waiting using sleep)
   # Once that result is created we move it to siimon5
-  scp -P 2235 $base_path/$dest/${name}.tif manuelsuarez@siimon5.cimat.mx:~/data/cimat/dataset-${dataset}/$dest/${name}.tif
+  zip $base_path/$dest/WindField_Point.zip WindField_Point.*
+  scp -P 2235 $base_path/$dest/WindField_Point.zip manuelsuarez@siimon5.cimat.mx:~/data/cimat/dataset-${dataset}/$dest/${name}_windfield_point.zip
   #scp -r -P 2235 $base_path/$dest/$name/${name}_VH.tif manuelsuarez@siimon5.cimat.mx:~/data/cimat/dataset-noaa/sentinel1/wind/${name}_VH.tif
   # Remove temporary files 01-05
   #
@@ -109,9 +115,9 @@ for fname in $(cat $base_path/${source}_list${num}.txt); do
   if test -f $base_path/$temp/$name.tif; then
     rm $base_path/$temp/$name.tif
   fi
-  # Remove tiff files
+  # Remove shp files
   if test -f $base_path/$dest/$name.tif; then
-    rm $base_path/$dest/$name.tif
+    rm $base_path/$dest/${name}_windfield_point.tif
   fi
   # Remove unziped product
   if test -d $base_path/$source/$name.SAFE; then
